+++ org.rssowl.ui/src/org/rssowl/ui/internal/dialogs/properties/Messages.java
@@ -66,7 +66,9 @@ public class Messages extends NLS {
   public static String GeneralPropertyPage_FORCE_UPDATE_FOLDER;
   public static String GeneralPropertyPage_MANAGED_BY_ANCESTOR;
   public static String GeneralPropertyPage_OVERRIDE_PROXY_FOLDER;
-  public static String GeneralPropertyPage_BYPASS_PROXY;
+  public static String GeneralPropertyPage_PROXY_MODE_LABEL;
+  public static String GeneralPropertyPage_PROXY_MODE_USE_PROXY;
+  public static String GeneralPropertyPage_PROXY_MODE_BYPASS;
   public static String GeneralPropertyPage_WARNING;
   public static String InformationPropertyPage_BOOKMARKS;
   public static String InformationPropertyPage_CREATED;